Add step to block traffic to the Neutron API while ovn-db-sync runs
It is recommended to block traffic to the Neutron API when
neutron-ovn-db-sync-util is running in the "repair" mode so that no new
changes in the Neutron API are made in the same time.
This isn't really critical in the lab environment but as it is good
practice in the production environments, lets add such steps to the
training's guide too.